ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
redis==5.0.1
pymysql==1.1.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
joblib==1.3.2
python-multipart==0.0.6
python-dotenv==1.0.1
pytest==8.3.2
httpx==0.27.0